+++ packages/backend/eslint.config.js
@@ -19,7 +19,7 @@ export default [
 		languageOptions: {
 			parserOptions: {
 				parser: tsParser,
-				project: ['./tsconfig.json', './test/tsconfig.json', './test-federation/tsconfig.json'],
+				project: ['./tsconfig.json', './scripts/tsconfig.json', './test/tsconfig.json', './test-federation/tsconfig.json'],
 				sourceType: 'module',
 				tsconfigRootDir: import.meta.dirname,
 			},